requirements.txt
langchain
langchain_community
chainlit==1.0.506
langchain-groq
groq
langchain-openai
python-dotenv
huggingface_hub
chromadb
openai
torch
pymupdf4llm
langchain_nvidia_ai_endpoints
transformers
pytesseract
tesseract
pdf2image
unstructured[pdf]
opencv-python-headless
layoutparser
pi-heif
pypdf
matplotlib
unstructured_inference
langchain-google-genai
ipython